Basic join performance tests
