+++ build.sbt
@@ -3,22 +3,13 @@
 
 scalaVersion := "2.10.4"
 
-sparkVersion := "1.3.0"
-
 sparkPackageName := "databricks/spark-sql-perf"
 
-// Don't forget to set the version
-version := "0.0.1-SNAPSHOT"
+version := "0.0.4-SNAPSHOT"
 
 // All Spark Packages need a license
 licenses := Seq("Apache-2.0" -> url("http://opensource.org/licenses/Apache-2.0"))
 
+sparkVersion := "1.4.0"
 
-// Add Spark components this package depends on, e.g, "mllib", ....
 sparkComponents ++= Seq("sql", "hive")
-
-// uncomment and change the value below to change the directory where your zip artifact will be created
-// spDistDirectory := target.value
-
-// add any sparkPackageDependencies using sparkPackageDependencies.
-// e.g. sparkPackageDependencies += "databricks/spark-avro:0.1"
